method loadDatabase added

+++ ex00/inc/BitcoinExchange.hpp
@@ -6,7 +6,7 @@
 /*   By: ghambrec <ghambrec@student.42heilbronn.    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/02/20 13:42:14 by ghambrec          #+#    #+#             */
-/*   Updated: 2026/02/21 12:35:18 by ghambrec         ###   ########.fr       */
+/*   Updated: 2026/02/23 13:13:41 by ghambrec         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,6 +16,7 @@
 #include <string>
 #include <map>
 #include <regex>
+#include <fstream>
 
 inline const std::string DB_FILE{"data/data.csv"};
 inline const std::regex REGEX{R"(...)"};
@@ -23,7 +24,7 @@ inline const std::regex REGEX{R"(...)"};
 class BitcoinExchange
 {
 	private:
-		std::map<std::string, float> btc_;
+		std::map<std::string, double> btc_;
 
 	public:
 		BitcoinExchange();
